chore: more parsing from datapack

+++ crates/pumpkin-world/src/generation/block_state_provider.rs
@@ -234,10 +234,10 @@ impl DualNoiseBlockStateProvider {
 
     fn get_slow_noise(&self, x: f64, y: f64, z: f64, sampler: &DoublePerlinNoiseSampler) -> f64 {
         sampler.sample(
-            x * self.slow_scale,
-            y * self.slow_scale,
-            z * self.slow_scale,
-        )
+            (x * self.slow_scale) as f32,
+            (y * self.slow_scale) as f32,
+            (z * self.slow_scale) as f32,
+        ) as f64
     }
 }
 
@@ -281,10 +281,10 @@ impl NoiseBlockStateProviderBase {
             false,
         );
         sampler.sample(
-            pos.0.x as f64 * self.scale as f64,
-            pos.0.y as f64 * self.scale as f64,
-            pos.0.z as f64 * self.scale as f64,
-        )
+            pos.0.x as f32 * self.scale,
+            pos.0.y as f32 * self.scale,
+            pos.0.z as f32 * self.scale,
+        ) as f64
     }
 }
