+++ crates/pumpkin-world/src/generation/noise/router/density_function/mod.rs
@@ -24,13 +24,13 @@ pub trait IndexToNoisePos {
 }
 
 pub trait NoiseFunctionComponentRange {
-    fn min(&self) -> f64;
-    fn max(&self) -> f64;
+    fn min(&self) -> f32;
+    fn max(&self) -> f32;
 }
 
 pub trait StaticIndependentChunkNoiseFunctionComponentImpl: NoiseFunctionComponentRange {
-    fn sample(&self, pos: &Vector3<i32>) -> f64;
-    fn fill(&self, array: &mut [f64], mapper: &impl IndexToNoisePos) {
+    fn sample(&self, pos: &Vector3<i32>) -> f32;
+    fn fill(&self, array: &mut [f32], mapper: &impl IndexToNoisePos) {
         array.iter_mut().enumerate().for_each(|(index, value)| {
             let pos = mapper.at(index, None);
             *value = self.sample(&pos);
@@ -41,8 +41,8 @@ pub trait StaticIndependentChunkNoiseFunctionComponentImpl: NoiseFunctionCompone
 pub struct Wrapper {
     pub input_index: usize,
     pub wrapper_type: WrapperType,
-    min_value: f64,
-    max_value: f64,
+    min_value: f32,
+    max_value: f32,
 }
 
 impl Wrapper {
@@ -50,8 +50,8 @@ impl Wrapper {
     pub const fn new(
         input_index: usize,
         wrapper_type: WrapperType,
-        min_value: f64,
-        max_value: f64,
+        min_value: f32,
+        max_value: f32,
     ) -> Self {
         Self {
             input_index,
@@ -64,12 +64,12 @@ impl Wrapper {
 
 impl NoiseFunctionComponentRange for Wrapper {
     #[inline]
-    fn min(&self) -> f64 {
+    fn min(&self) -> f32 {
         self.min_value
     }
 
     #[inline]
-    fn max(&self) -> f64 {
+    fn max(&self) -> f32 {
         self.max_value
     }
 }
@@ -77,13 +77,13 @@ impl NoiseFunctionComponentRange for Wrapper {
 #[derive(Clone)]
 pub struct PassThrough {
     input_index: usize,
-    min_value: f64,
-    max_value: f64,
+    min_value: f32,
+    max_value: f32,
 }
 
 impl PassThrough {
     #[must_use]
-    pub const fn new(input_index: usize, min_value: f64, max_value: f64) -> Self {
+    pub const fn new(input_index: usize, min_value: f32, max_value: f32) -> Self {
         Self {
             input_index,
             min_value,
@@ -99,12 +99,12 @@ impl PassThrough {
 
 impl NoiseFunctionComponentRange for PassThrough {
     #[inline]
-    fn min(&self) -> f64 {
+    fn min(&self) -> f32 {
         self.min_value
     }
 
     #[inline]
-    fn max(&self) -> f64 {
+    fn max(&self) -> f32 {
         self.max_value
     }
 }
